Refactor: Move the WorldServer prototype around
It doesn't do much, but I'll probably want to research the various game mechanics more in the future. So for the time being, the server can stay, but it should be moved out of the way a little.
