+++ start-server.lua
@@ -1,0 +1,5 @@
+package.path = "?.lua"
+
+local WorldServer = require("Core.WorldServer.WorldServer")
+
+WorldServer:Start()
